Fix an assertion failure when starting an async transaction from within a sync transaction
Starting an async write transaction from within a write transaction while the
async state was idle would result in the background thread waiting on the write
lock while the Transaction already held it on the current thread. This isn't an
inherently invalid thing to do, but it's something that can't happen in any
other way and had a pile of race conditions. Rather than try to get this to
work correctly, it's easier to just schedule the async write to be kicked off
when the synchronous write is committed, which makes this scenario work much
more like other scenarios.