Set the linked frameworks on the correct SPM target (#5254)
SPM will build library targets as either static or dynamic depending on how
many targets depend on them and if building as static would result in
duplication. When the core library happens to be built as static, it doesn't
use the linker settings and defining them in the downstream consumers works.
However, when it's built as dynamic it actually does need to link all the
libraries it uses.