Ensure scheduler is not deleted until callback has been done
This is fixing a problem introduced in 14210d1dbd235

The issue was discovered in realm-js, where async open would
hang because the scheduler was deleted before the callback was done on the
main thread. Beforehand the EventLoopDispatcher has a State member that
would retain the scheduler until the invocations were done.