Improve handling of keychain errors on macOS
While on iOS the keychain basically just always works, on macOS the keychain is
a user-managed thing with permissions and accessing it can either fail or pop
up a prompt for the user to approve or reject it, and we need to not just crash
when that happens.

If we can't access the keychain, this makes it so that we simply store the
metadata realm unencrypted.