Prevent advancing to intermediate versions when advancing notifications
We would like to avoid to advance to versions that are not currently "alive",
meaning that they have at least one read transaction attached to it. This will
make it easier to free space accupied by "dead" versions.

We can instead start a transaction at the lastest version  and use the
history information here to advance the notifiers. The history
contains the changesets from the oldest referenced version up to the current..