Decouple baas test and test sync server test initialization
Using TestSyncManager for baas tests meant that we were starting the sync
server for each of those tests. In addition to the small performance hit, this
made it possible to accidentally use the wrong server. This also includes some
incidental refactoring to tests where it turned out that TestSyncManager's
abstractions no longer fit with what the tests actually wanted.