Adapt AppError to new Exception framework
AppError is no longer based on std::error code, which means that the various places where
we beforehand used std::error_code as a parameter (notably callbacks) had to change.
We now use Status as the standard way of passing status around (which seems logical).
However, it is nol only AppError information we pass around, but also information from
SyncError. SyncError is heavily based on std::error_code and it would be a rather big
task to step away from that. So in order to be able to also pass std::error_code around
in a Status object, Status class is enhanced to also to contain that information. This
allows us to make SyncError inherit from SystemError with the information stored in the
underlying Status object.