Make thread-safe-ref transferable
The callback function called during realm_async_open_task_start must
receive a pointer to a thread-safe-ref that can be transferred to another thread.
For that to be possible the object must be allocated on the heap so that the
callee owns the object. Previously the pointer was pointing to a stack based
local variable.