Make MemRef trivially destructible
This makes it so that it is passed in registers rather than on the stack (on
Itanium ABI platforms), which gives a small but measurable performance
improvement to basically everything.

A side effect of this is that Obj becomes trivially destructible as well, which
results in clang now warning about unused Obj variables.