Fix a data race on write_validator_factory
PendingBootstrapStore is created early in session initialization, which
performs a write and thus reads the write validator. This can race with setting
the write validator on the thread which created the sync session, and tsan
complains about it.

Moving the initialization to SyncSession's constructor ensures that it's set
before we create the `sync::Session`.