Fix a data race in the write mutex ticketing system
The emulated implementation of InterprocessCondVar requires that the associated
mutex be held when `notify_all()` is called because it performs non-atomic
reads and writes on the shared part, which would be a data race if done without
a mutex covering the acceses. This is already noted as a precondition of the
function, which DB::do_end_write() violated.