Fix a data race while writing audit events
Calling `DB::set_replication()` inside a write transaction is _almost_ safe to
do, but it turns out that there's a few scenarios where the sync client will
call `DB::get_replication()` outside of a write. Scoping the replication
setting to just the specific Table fixes this (as live Table accessors actually
are thread-confined) and also simplifies the logic a bit.