Fix a race condition when recording audit write events
Recording audit write events requires having access to the old and new versions
of the objects so that we can report the old value of properties in the events.
This was done by beginning a read Transaction at the old version after
committing the write transaction, but this is too late; if another write
transaction happens on a different thread (say, by the sync client) in between
the commit and the call to `record_write()` the old version may be cleaned up
already.