Skip GlobalKey fixups when it's not needed
Rewriting the changesets after receiving the client file ident is only required
for tables without primary keys, which the new server doesn't support. We need
to keep the functionality for the sake of the existing sync tests, but we
don't want to do all that work in the normal usage.