Handle collection notifiers not being ready in a QoS-friendly way
On Apple platforms, waiting on a condition variable does not support donating
the waiting thread's priority to the thread doing the work, as the OS doesn't
know what thread will notify the CV. This can result in stalls from QoS
inversions where a high priority thread is waiting on a low priority thread to
complete work, but the low priority thread doesn't get scheduled.

Instead of using a condition variable, we now guard `run_async_notifiers()`
with a mutex. If notifiers aren't ready, we try to acquire the mutex on the
advancing thread. If the background thread is actively running the notifiers,
this'll block and donate the advancing thread's priority and time slice to the
background thread. If it's not, then we just call `run_async_notifiers()` on
the current thread.

Updating the tests to work with this revealed something that was working as
designed, but contrary to what all of our SDK documentation says because the
design was weird and bad. Previously, `refresh()` did not always actually
advance to the most recent version. Instead, if there was a newer version which
notifiers were ready for it would advance to that even if it's not the latest
version. If it couldn't advance without blocking, then it would block until the
most recent version was ready.

This behavior made `refresh()` mostly useless as it didn't guarantee that
committing a write transaction on one thread and then refreshing on another
would make the second thread see the write, but it also wasn't a best-effort
advance without blocking (which is done via `notify()`). Removing this
behaviour is arguably a breaking change, but it's aligning the actual behavior
with the documented behavior so I'm calling it a bug fix.