Greatly improve performance of sorting dictionaries (#5168)
Dictionary lookup by index is slow, so read all of the keys/values from the
dictionary into a vector and sort that instead. Using a dictionary iterator to
read the values replaces O(N log N) ClusterTree lookups with O(log N), so this
is asymptotically faster. The benchmark is sped up from 3.77 seconds to 21.62
milliseconds, but more reasonably sized Dictionaries will see proportionally
smaller benefits.

The downside of this is that the array of Mixed requires 24 bytes of scratch
space per element in the dictionary. We already require 8 bytes per element to
store the results, so this is just a constant factor increase rather than an
aysmtotic change in memory use and it's probably not a problem.