Add stricter checks for valid reads on encrypted files
If a read on an encrypted file failed for any reason other than the hmac check
failing, we simply ignored the failure and skipped the read. This is required
for when we're only reading a page so that we can subsequently write to it and
is almost okay the rest of the time, but it resulted in misleading errors in
some cases.