Fix a use-after-free in client reset when the original RealmCoordinator is gone
`DB::get_encryption_key()` returns a pointer to a buffer owned by the thing
which created the DB, which may no longer be valid when we're downloading the
new Realm for client reset.

This was hit by the "callbacks are seeded with Realm instances even if the
coordinator dies" client reset test when running with encrypt all enabled.