Fix assertion failures after calling SectionedResults::reset_section_callback()
If the SectionedResults had been previously evaluated, reset_section_callback()
would leave it in an inconsistent state where
m_has_performed_initial_evalutation was false but the previous run info was
non-empty. #5926 introduced some assertions that now fail when this happens,
but I suspect the actual behavior was also broken before that.