Fix corruption of freelist during compaction (#6054)
Wrong value for the size of the top array was used because values were added after the sizes were calculated. Now the decision to compact the realm is moved to a point before the sizes are calculated so that the correct top array size is used. The reason for moving the logic to the point after the calculation of the sizes was to use an updated values for the free space size - mostly in order to get a test passing, but that was obviously a wrong decision.