Trim algorithm for determining evacuation point (#6060)
This will enable a test that was previously disabled.
The problem with the test was that the compacted realm was still bigger than expected. This was because the free space size used in the algorithm was way smaller than what was actually free, so the used space size was way bigger, which resulted in the evacuation point being too high.