Fix BadVersion exceptions when using notifications and sequential writes on a background thread
Abiding by v13's new rules for version liveness requires pinning versions used
for notifications in more places.

Previously the notifiers relied on the logic that all VersionIDs greater than
the oldest live one were valid, which means that there was no need to pin
things like the version we were going to advance a Realm to, as it was always
greater than the current version of the Realm.

To fix this, most of the places where we were storing VersionIDs in
notification stuff now store Transactions instead.

All these additional Transactions being created appears to have a negligable
performance impact. The additional readlock increments and decrements are very
not free, but the Transaction objects wrapping them are minor in comparison.