Optimize acquiring and releasing read locks when one is already held (#6115)
* Optimze acquiring and releasing read locks when one is already held

Acquiring the file lock which guards the lock file is an expensive operation
which dominates the runtime of creating a new Transaction. To avoid doing that
when possible, this makes it so that we only update the lock file when
transitioning between zero and one of each lock type, and all other refcounting
is done purely in memory.

This replaces a file lock with a much cheaper local mutex whenever we already
have a read lock of that type for the given version (or when releasing, if we
still have a read lock of that type). This makes the temporary Transactions
used for things like notifications much cheaper.

Benchmark results:

TransactionDuplicate (v13.1.2): 10.52ms
TransactionDuplicate (this): 1.80ms

OS notification delivery (tg/notifier-badversion): 800ms in do_refresh()
OS notification delivery (this): 413ms in do_refresh()
OS notification delivery (v13.1.2): 277ms in do_refresh()

* attempt to silence warning on windows

Co-authored-by: Finn Schiermer Andersen <finn.schiermer.andersen@gmail.com>