Fix BadVersion exceptions when using notifications and sequential writes on a background thread (#6106)
* Fix BadVersion exceptions when using notifications and sequential writes on a background thread

Abiding by v13's new rules for version liveness requires pinning versions used
for notifications in more places.

Previously the notifiers relied on the logic that all VersionIDs greater than
the oldest live one were valid, which means that there was no need to pin
things like the version we were going to advance a Realm to, as it was always
greater than the current version of the Realm.

To fix this, most of the places where we were storing VersionIDs in
notification stuff now store Transactions instead.

All these additional Transactions being created appears to have a negligable
performance impact. The additional readlock increments and decrements are very
not free, but the Transaction objects wrapping them are minor in comparison.

* Optimize acquiring and releasing read locks when one is already held (#6115)

Acquiring the file lock which guards the lock file is an expensive operation
which dominates the runtime of creating a new Transaction. To avoid doing that
when possible, this makes it so that we only update the lock file when
transitioning between zero and one of each lock type, and all other refcounting
is done purely in memory.

This replaces a file lock with a much cheaper local mutex whenever we already
have a read lock of that type for the given version (or when releasing, if we
still have a read lock of that type). This makes the temporary Transactions
used for things like notifications much cheaper.

Benchmark results:

TransactionDuplicate (v13.1.2): 10.52ms
TransactionDuplicate (this): 1.80ms

OS notification delivery (tg/notifier-badversion): 800ms in do_refresh()
OS notification delivery (this): 413ms in do_refresh()
OS notification delivery (v13.1.2): 277ms in do_refresh()

Co-authored-by: Finn Schiermer Andersen <finn.schiermer.andersen@gmail.com>