Remove ability to synchronize objects without primary key
With the legacy server this was based on GlobalKey, but if we at some
point will allow objects without primary key to be synced to MongoDB
server, we will probably have to invent a new strategy.

As we remove support for GlobalKey we will have to change the way we
allocate ObjKeys for tombstones. Beforehand we used a hash of the
primary key to construct a GlobalKey and from that an ObjKey. Now we just
use the current ObjKey to construct a corresponding unresolved key. But
it has the effect that there is not an easy way to come from primary key
to potential unresolved key, so we will have to search for a potential
tombstone.