Make getting a complete view of the schema optional
Rather than tying this to the Additive schema mode, make it an option which can
be used with all applicable schema modes. This also fixes several bugs in the
implementation which would result in it failing to actually have a complete
view of the schema in some cases, and could result in frozen Realms having an
incorrect schema.

This also incidentally improves the performance of freezing Realms as part of
the fix for the schema management and fixes a few race conditions related to
cached frozen Realms.