Fix an invalid use of frozen Realms in client reset (#6230)
This turned out to not be the source of the problem I was looking in to, but
`get_realm(config, VersionID())` does strange and invalid things that appears
to work in simple cases but can cause weird problems. VersionID() is turned
into an actual version only when acquiring a read lock, and won't turn into the
same version every time for a given frozen Realm.