Actually sort in Set<Mixed>::sort()
Now that Set<Mixed>'s on-disk storage order doesn't match our sort order,
Set<Mixed>::sort() needs to actually sort the items rather than being a no-op.
It remains a no-op for other types of set.

Adjusting the data used for tests to test this also revealed that
Lst::distinct() picked an arbirary value of the equal ones rather than the
first. This isn't neccesarily incorrect, but it makes testing distinct more
complicated and can have surprising results when the collection isn't sorted.
Making it always use the first is simple and doesn't add any overhead, so it
now does that.