Delete some obsolete cruft (#5662)
The SSE/AVX code was written in 2012 and mostly hasn't been touched since then.
In the time since then, runtime checking for SSE4.2 support has become
pointless (CPUs supporting it were first released in 2008, so it was somewhat
recent in 2012 and now is not at all recent), and we never actually wrote
anything which requires AVX. There was also some SSE code which has been
disabled since 2012.

set_string_compare_method() and everyhing related to it has never actually been
used by any SDK, and is not really the correct solution to the problem anyway.

The encryption code used `ssize_t` in a few places because it was originally
written to use `pread` and friends, but the functions it now calls use `size_t`
and `ssize_t` doesn't exist on Windows. The use of ssize_t in the Windows
networking code was simply incorrect; the winsock functions return `int`.