Eliminate IncrementalChangeInfo
When the notification code was originally written, opening a new SharedGroup
was a very expensive operation and it was important to minimize the number of
them used. This lead to some complicated logic to handle multiple source
versions for notifiers where the single SharedGroup was incrementally advanced
between the versions. This is now all obsolete; newly registered notifiers hold
their own Transaction and can just be advanced separately.