(RJS bindgen) Removed simple test/example used to bootstrap bindgen project
Now that we can run the real tests, this file has outlived its usefulness. It
stopped working a month ago when we switched from tsm to tsx (because they
disagree about which files should be treated as cjs vs esm modules), and it
isn't worth the effort to make it work again.