Allow registering notifiers in write transactions prior to the first change (#6560)
This fixes one of the common problems resulting from
https://github.com/realm/realm-swift/issues/4818, as it makes it so that adding
new callbacks from within a notification callback will _usually_ work even if
the notification was triggered by beginning a write transaction. It still will
not work if a previously invoked callback made any changes (which would be a
very strange thing to do, but possible).