Replace StatusWith with Expected
C++23's std::expected fills the same role as StatusWith, but is better designed
and more flexible.

This implementation diverges from the standard by only supporting `Status` as
the error type and not allowing `Status` as the expected type, which eliminates
the need for the `unexpected` disambiguator. In addition, `Exception` is thrown
rather than `std::bad_expected_access`. It should otherwise be aligned with
`std::expected`.

This greatly simplifies `util::Future` due to that `Expected` takes care of
most of the complexity of storing `void`, eliminating the need for the
`FakeVoid` things.