Silense a Valgrind error
This error was introduced in 601805bd148aa8, where the destructor on
Mixed was removed. It was not possible to actually prove that Valgrind
was right in reporting a read from uninitialized memory, but changing
the logic a bit could make the error go away. The new code is
probably slightly slower, but it is not on the critical path.