Add an option to pass a live Realm to the before-reset callback
For complicated performance reasons, Cocoa doesn't pass the schema in the
RealmConfig and instead calls update_schema() later. This means that when a
client reset happens during an async open, the object store code doesn't have
access to the actual schema yet, and the write to update the schema needs to
happen in the SDK code.