Tweak embedded object checking loop to skip over objects already seen (#6803)
Say we are checking for cycles starting from object A and traverse our
way down through  A -> B -> C -> D -> E and A -> F -> C -> D -> E.
Prior to this change, we'd traverse every single property, but if we've
already checked C -> D -> E via object B, we can avoid checking it via
object F. Cumulatively, this can save a bunch of redundant checks

We also stop keeping track of all the object types we've seen. This is
because if there is a cycle A -> E -> B -> C -> D -> E, we can identify
this when E is the object being considered as the start point (seeing as
the caller goes through all embedded objects defined in the schema
anyway).

If it is detected when A is the object being considered, we're not really
adding any additional useful information. Instead, the do_check loop
becomes much simpler (and we avoid copying) if we check all child
objects are not embedding the 'start' object only