Fix handling of users with multiple identities
ROS users were identified by the pair (identity, auth_url), and in the v10 port
this was translated to (user_id, provider_type), but that isn't actually how
Atlas users work. An App is the equivalent of the old auth_url, and within an
App user_id uniquely identifies a user. Users don't actually have a
"provider type" at all: users have one or more identities, each of which has a
provider type, and the same SyncUser should be used regardless of which
identity was used to log in.

This had surprisingly mild consequences, but was visibly broken in a few ways.
Logging into the same user using two different identities resulted in two
SyncUsers with the same user id. Opening the same partition (or any flx Realm)
with both users resulted in both using whichever SyncUser happened to open it
first, and the second would not create a SyncSession. This meant that most of
the potential bad things (such as creating two session for one file) didn't
actually happen, but the second user's list of sessions would be "incorrect",
and removing one of the users would remove the incorrect set of local files.

Linking identities to anonymous users resulted in the user still being treated
as anonymous. The primary negative effect of this was that linking an identity,
logging out, then logging back in would result in the user being removed
entirely in between, forcing the re-download of all Realms (and the loss of any
un-uploaded data).

This bumps the schema version of the metadata Realm primarily for the sake of
recovery on metadata Realms in invalid states: the migration block handles the
case of multiple users with the same user id and unifies them. Since this needs
a migration anyway, it fixes some incidental problems with the metadata Realm's
schema which weren't fixable without a migration: the marked_for_deletion
column was redundant with the Removed state, identity objects were leaked
due to not being embedded, and the local_uuid field is no longer used for
anything other than opening files written by early v10 versions, so there's no
need to populate it for new users.