Add a test for TLS error reporting and fix some problems
With SecureTransport only a single error code was reported as the TLS handshake
failing, but actually every error code with the secure transport error category
indicates that.

With OpenSSL things were worse: it had a specific list of error messages that
indicate a TLS error. This didn't actually work, as the error message wasn't
being populated correctly and TLS errors would end up with errors like "RSA lib"
rather than what the code was checking for (and these useless errors were
forwarded on to the SDK).