Add an evergreen job to validate that all installed headers build (#6954)
We've had a few issues with headers being added as dependencies of an installed
header but not being installed, so add a CI job that validates that all
installed headers can be compiled with only the other installed headers
available. This also catches some incidental missing header inclusions where a
header failes to include everything it depends on.

Only runs on macOS since Cocoa seems to be the only SDK which relies on this
being valid.