Simplify intra-component commit notification
The concrete bug which this fixes is that `await realm.subscriptions.update {};
await realm.refresh()` would hang forever. SubscriptionStore writes failed to
notify RealmCoordinator of the write, so the async refresh would see that the
Realm is not on the latest version, register a handler to be called when the
autorefresh happened, and then nothing would ever schedule the autorefresh.

The sync client needs to be notified of non-sync writes and notify non-sync
components when it performs writes. When it was first written, DB did not exist
yet and so this was orchestrated via RealmCoordinator. However, that's a very
awkward place to do it: not all writes go via RealmCoordinator, and the
lifetime of sync sessions isn't actually tied to a coordinator. Nowadays we do
have DB, and handling commit notifications there greatly simplifies everything.

There was also a *second* mechanism for notifying the sync client of writes
which modified the subscription store. This appears to have been entirely
redundant and unnecessary.