Simplify Session's tracking of which versions should be uploaded
The documentation suggests there was once a mechanism for uploading up to a
specific version and then stopping, but this is now only used for sending QUERY
messages at the correct time, and that can be done more directly. This cuts
down on the amount of state that needs to be tracked and sometimes (very
insignificantly) improves upload latency.