Simplify sync::Transformer and eliminate the abstract interface
Transformer objects used to have some local state which had to be stored
between uses and had the implementation hidden in the cpp file. Neither of
those are true any more, so the design doesn't make sense either. It can now
just be a stack-allocated local created when needed rather than something which
has to be stored and reused.