Rewrite the local changesets in-place for client reset recovery
Rather than discarding the old sync history and creating new commits, update
the individual changesets which are being recovered in-place while keeping the
original version numbers. This significantly simplifies recovery when pending
subscriptions are involved, as they no longer need to be updated at all, and it
lets us apply the client reset in a single write transaction, which improves
notifications and avoids a lot of problems if another thread has the Realm open
while the client reset is happening.

It also avoids causing problems for the server, which doesn't like overly large
changesets, and sometimes ran into problems due to all of the pending local
changesets being merged into one.