Fix a BadVersion exception from initial_subscriptions plus a client reset on an async open
If the very first open of a flexible sync Realm triggered a client reset, the
configuration had an initial subscriptions callback, both before and after
reset callbacks, and the initial subscription callback began a read transaction
without ending it (which is normally going to be the case), opening the frozen
Realm for the after reset callback would trigger a BadVersion exception.

The commit to create the initial subscriptions resulted in the `before` Realm
being one version of out date, so once it was closed and the read lock released
that VersionID was no longer valid and could not be used to obtain new
transactions.