Improve performance when resurrecting an object with many backlinks (#7218)
The process of first removing the backlink from the tombstone and then
creating it again in the live object took a long time. As the backlinks
will end up being identical, we can just move the whole structure from
the tombstone to the new object and then just replace the link in
the linking object without worring about backlinks.

This change is only done for link and linklist properties.