Simplify base url updating
We already have a more robust implementation of Uri splitting that should be
used instead of a second implementation.

Putting the redirect count on Request is more complicated than passing it
directly, and testing that there are not unexpected additional requests made
via checking the request count directly caught a bug which validating the
redirect count did not.

A few places used std::optional<std::string> where none and empty were treated
identically, which means that it should be just an unwrapped string.

The wrapper to heap-copy Requests did not do anything useful.