improve tests and defensivly handle the same case
The changes to handle the context flag propagation in this commit
are not actually fixes because the context flag is not used for
any other type than BPTree<ObjKey> (LnkLst), and these
other cases are for split_root() which is only ever called for
BPTree<Int> in the case of backlinks storage which doesn't care
about the context flag. So these are just defensive in case future
code does rely on this.