Improve performance with very large numbers of notifiers
With large notifier counts the lock contention on CollectionNotifier::m_realm
becomes a very significant overhead, with unlock() taking upwards of 50% of the
total runtime. This is only a problem when the notifiers run on the background
thread and the target thread forces a refresh, as there's otherwise no
contention. Using an atomic variable to avoid the mutex solves the problem
entirely.