Make Scheduler's factory function a function pointer rather than a UniqueFunction (#7467)
This makes it trivially destructible and so eliminates a spurious crash related
to reading the variable on one thread while it's being destroyed on another
(due to the application exiting). This breaks the C API for registering a
custom default scheduler, but that turns out to be unused and so it can be
deleted. All users of the C++ scheduler API work with just a function pointer.