Update the parent's content version when bumping it in a nested collection (#7470)
The parent of a nested collection will always be up to date after making a
modification to the collection as we update it prior to performing the
modification and the modification cannot make the parent stale. Failing to
update the parent's content version meant that the parent was always considered
stale after an update to a child collection, resulting in the child getting
updated between each modification even when the modifications were all done via
a single accessor.