Truncate the metadata realm in-place rather than deleting and recreating it (#7526)
This solves some locking problems in multi-process scenarios. The flow of
checking if the key is valid, deleting the file if not, and then recreating the
file if needed needs to be done with an exclusive lock on the file held. This
can't be done with DB::call_with_lock() because there isn't a good way to
initialize the new file from within the call_with_lock() callback, and instead
needs to be pushed into the file initialization guarded by the control mutex.

Co-authored-by: Finn Schiermer Andersen <finn.schiermer.andersen@gmail.com>