Eliminate all copies of Request
We sometimes need to retry requests after doing things like fetching a new
access token, so the response handlers need access to the request. This
requires either copying the request or heap allocating it, and as the request
body can be quite large heap allocating it is better. Previously we created
them on the stack and then moved them to the heap, but as every request goes
through the path where they're moved to the heap they should just start there.