Simplify the SessionWrapper lifecycle a bit
Initializing a sync::Session was a multi-step process of creating the Session
with a config, configuring some addition things via mutation functions, and
then calling bind(). We can simplify this a bit by pushing everything into the
config struct and binding inside the wrapper's constructor, eliminating the
constructed-but-not-initialized state and letting us make more of the members
const.