Fix compilation with `swift build`
Adding a resource to a target makes `swift build` try to force-include
Foundation.h when building source files in that target, which breaks if any of
the source files aren't obj-c. To work around this, the resources need to be in
a separate target from the source files.

xcodebuild breaks if a target has zero source files, so add an empty obj-c file
to build as part of this new resource target.