More robust support for reading an encrypted Realm
without assuming that the page size is the same
across devices. In particular, we now count reading
blocks of 0's as a valid block rather than returning
early. This supports data patterns of zero'd blocks
between two valid data blocks instead of always assuming
that the zero'd blocks are at the end.