Don't emit transaction log instructions for mutations on newly-created objects
The non-sync transaction logs are only used to drive notifications and
notifications don't care about mutations on objects in the same commit as the
objects were created it, so we don't need to emit the instructions at all. This
significantly cuts the size of the transaction log for commits which are
primarily inserting objects.

This does a very basic check for "newly-created" which tracks the most recently
created object for each table and skips mutation instructions for that object.
This handles recursively creating an object and all of its embedded objects
without the overhead of tracking every single object created within a
transaction, and insertion workflows will typically not return to an object
after creating another object in the same table.

This requires adding an additional small amount of tracking for embedded
objects, as Replication previously didn't know when new embedded objects were
created.