Fix race in schema migration tests (#7766)
Some of the schema migration tests test the recovery of subscriptions and objects after a schema migration. This requires that the subscriptions and objects are committed when the synchronization session is paused so nothing is uploaded. `SyncSession::pause()` is used for this, but pause is an async operation so there is a race between pausing the session and uploading the data and subscriptions. `SyncSession::shutdown_and_wait()` is used instead.