Fix compacting a Realm file using the existing encryption key
`compact()` stored the existing encryption key in a `std::string`, which
expects a nul-terminated string, not a fixed-size buffer. If the key contained
any nul bytes then the key would be truncated and garbage data would be used as
the encryption key instead, and if it didn't then arbitrarily large additional
amounts of memory would also be copied into the buffer. The tests happened to
always use a nul-terminated string as the key and so worked by coincidence.